requirements.txt
requests
beautifulsoup4
PyYAML
python-dotenv